ENH: Add Abstract interface for marker finder
