+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode/*
 venv/*
 *.pyc
+.venv/*
