+++ requirements.txt
@@ -1,8 +1,9 @@
-fastapi
-fastapi-utils
+fastapi==0.100.0
+fastapi-utils==0.2.1
 hjson
 opencv-contrib-python==4.5.5.64
 pydantic~=1.10.15
+scipy
 uvicorn[standard]
 websocket
 websockets
