+++ src/common/structures/__init__.py
@@ -1,4 +1,5 @@
 from .capture_format import CaptureFormat
+from .capture_status import CaptureStatus
 from .component_connection_static import ComponentConnectionStatic
 from .component_role_label import \
     ComponentRoleLabel, \
@@ -15,7 +16,6 @@ from .corner_refinement import \
     CORNER_REFINEMENT_METHOD_DICTIONARY_TEXT_TO_INT
 from .detection_parameters import DetectionParameters
 from .detector_resolution import DetectorResolution
-from .detector_status import DetectorStatus
 from .image_resolution import ImageResolution
 from .intrinsic_calibration import \
     IntrinsicCalibration, \
@@ -24,6 +24,7 @@ from .intrinsic_parameters import IntrinsicParameters
 from .marker_corner_image_point import MarkerCornerImagePoint
 from .marker_definition import MarkerDefinition
 from .marker_snapshot import MarkerSnapshot
+from .marker_status import MarkerStatus
 from .matrix4x4 import Matrix4x4
 from .pose import Pose
 from .pose_solver_status import PoseSolverStatus
