+++ src/detector/detector_app.py
@@ -24,6 +24,8 @@ import hjson
 import logging
 import os
 
+from src.detector.implementations import AbstractMarkerInterface, ArucoMarker
+
 
 logger = logging.getLogger(__name__)
 
@@ -31,11 +33,18 @@ logger = logging.getLogger(__name__)
 def create_app() -> FastAPI:
     detector_configuration_filepath: str = os.path.join(os.path.dirname(__file__), "..", "..", "data", "config.json")
     detector_configuration: DetectorConfiguration
+    #camera_interface: AbstractCameraInterface
+    marker_interface: AbstractMarkerInterface
+
     with open(detector_configuration_filepath, 'r') as infile:
         detector_configuration_file_contents: str = infile.read()
         detector_configuration_dict = hjson.loads(detector_configuration_file_contents)
         detector_configuration = DetectorConfiguration(**detector_configuration_dict)
-    detector = Detector(detector_configuration=detector_configuration)
+
+    #camera_interface = Detector.USBWebcamWithOpenCV()
+    marker_interface = ArucoMarker()
+    
+    detector = Detector(detector_configuration=detector_configuration,marker_interface=marker_interface)
     detector_app = FastAPI()
 
     # CORS Middleware
